refactored images upload
